Refactor to allow passing client name to auth provider
